Move tree_struct into features and simplify
Removes the render_tree function and its dependencies.
