+++ fonduer/features/content_features.py
@@ -21,7 +21,7 @@ from fonduer.supervision.lf_helpers import (
     tokens_to_ngrams,
 )
 from fonduer.models import TemporarySpan
-from fonduer.tree_structs import corenlp_to_xmltree
+from fonduer.features.tree_structs import corenlp_to_xmltree
 from fonduer.utils import get_as_dict
 
 DEF_VALUE = 1
